refactor(loop): split over-cap statusline.py into cohesive sibling modules
statusline.py was 681 LOC, over the 500 module-health cap (grandfathered by
the ratchet). Pure-move split by concern, mirroring the dispatch.py split:

* statusline_palette.py — shared ANSI palette, zone colors, the CSI / OSC 8 /
  overlay-prefix regexes, and the time / recency constants.
* statusline_render.py — presentation: StatuslineEntry / StatuslineZones, the
  colour decision, item / hyperlink formatting, the grouped-by-overlay render
  writer, and the Slack-mrkdwn reader.
* statusline_loops.py — the dedicated loop-line dashboard: live_loops_anchor
  and its DB-read seams (live leases, cadence resolution, mini-loop schedules,
  availability, pending questions) plus the recency-colour / next-tick
  formatters.

statusline.py is now a thin re-export of the public surface with an explicit
__all__ and identity preserved (statusline.render is statusline_render.render).
Every moved symbol is byte-identical bar deliberately re-pointed docstring
cross-refs. Importers of moved private symbols re-point to the canonical
sibling. Monkeypatch seam strings for the moved DB-read helpers re-point to
statusline_loops so the patches still affect live_loops_anchor behaviour.

statusline.py: 681 -> 56 LOC. statusline_loops.py 454, statusline_render.py
181, statusline_palette.py 52 — all under cap.
