+++ tests/teatree_loop/test_statusline_loop_owner_first.py
@@ -25,10 +25,10 @@ class TestLoopRunningTokenDropped:
     def test_line_leads_with_tick_chunk_not_loop_running(self) -> None:
         acquired_at = datetime.now(UTC) - timedelta(seconds=120)
         with (
-            patch("teatree.loop.statusline._live_loop_leases", return_value=[("loop-tick", acquired_at)]),
-            patch("teatree.loop.statusline._cadence_for_loop", return_value=720),
-            patch("teatree.loop.statusline._availability_segment", return_value=""),
-            patch("teatree.loop.statusline._pending_questions", return_value=0),
+            patch("teatree.loop.statusline_loops._live_loop_leases", return_value=[("loop-tick", acquired_at)]),
+            patch("teatree.loop.statusline_loops._cadence_for_loop", return_value=720),
+            patch("teatree.loop.statusline_loops._availability_segment", return_value=""),
+            patch("teatree.loop.statusline_loops._pending_questions", return_value=0),
         ):
             lines = live_loops_anchor()
         assert len(lines) == 1, repr(lines)
@@ -42,13 +42,13 @@ class TestLoopRunningTokenDropped:
     def test_mini_loop_only_line_has_no_loop_running(self) -> None:
         now = datetime.now(UTC)
         with (
-            patch("teatree.loop.statusline._live_loop_leases", return_value=[]),
+            patch("teatree.loop.statusline_loops._live_loop_leases", return_value=[]),
             patch(
-                "teatree.loop.statusline._mini_loop_schedules",
+                "teatree.loop.statusline_loops._mini_loop_schedules",
                 return_value=[("dispatch", now + timedelta(seconds=120), 600)],
             ),
-            patch("teatree.loop.statusline._availability_segment", return_value=""),
-            patch("teatree.loop.statusline._pending_questions", return_value=0),
+            patch("teatree.loop.statusline_loops._availability_segment", return_value=""),
+            patch("teatree.loop.statusline_loops._pending_questions", return_value=0),
         ):
             lines = live_loops_anchor()
         assert lines == ["dispatch 2m"], lines
@@ -56,17 +56,17 @@ class TestLoopRunningTokenDropped:
     def test_waiting_clause_still_appended(self) -> None:
         acquired_at = datetime.now(UTC) - timedelta(seconds=120)
         with (
-            patch("teatree.loop.statusline._live_loop_leases", return_value=[("loop-tick", acquired_at)]),
-            patch("teatree.loop.statusline._cadence_for_loop", return_value=720),
-            patch("teatree.loop.statusline._availability_segment", return_value=""),
-            patch("teatree.loop.statusline._pending_questions", return_value=2),
+            patch("teatree.loop.statusline_loops._live_loop_leases", return_value=[("loop-tick", acquired_at)]),
+            patch("teatree.loop.statusline_loops._cadence_for_loop", return_value=720),
+            patch("teatree.loop.statusline_loops._availability_segment", return_value=""),
+            patch("teatree.loop.statusline_loops._pending_questions", return_value=2),
         ):
             lines = live_loops_anchor()
         assert lines == ["tick 10m · waiting: 2 questions"], lines
 
     def test_still_empty_when_no_loops_live(self) -> None:
         with (
-            patch("teatree.loop.statusline._live_loop_leases", return_value=[]),
-            patch("teatree.loop.statusline._mini_loop_schedules", return_value=[]),
+            patch("teatree.loop.statusline_loops._live_loop_leases", return_value=[]),
+            patch("teatree.loop.statusline_loops._mini_loop_schedules", return_value=[]),
         ):
             assert live_loops_anchor() == []
